-change delay of the AI
-chaged stats to double (but added cast to int everywhere)

+++ ForgottenSchism/ForgottenSchism/world/Caster.cs
@@ -27,7 +27,7 @@ namespace ForgottenSchism.world
 
             stats.mana -= sp.ManaCost;
 
-            return sp.Damage + stats.traits.intel;
+            return (int)(sp.Damage + stats.traits.intel);
         }
 
         public String attack(Character c, Spell sp)
